feat: 一键场景编辑与其他场景编辑共用一个页面,TODO:删除automation-edit-yijian页面,优化automation-add代码

+++ src/package-automation/automation-add/index.ts
@@ -5,7 +5,14 @@ import pageBehavior from '../../behaviors/pageBehaviors'
 import { ComponentWithComputed } from 'miniprogram-computed'
 import { deviceStore, sceneStore, homeStore, autosceneStore, roomStore } from '../../store/index'
 import { PRO_TYPE, SENSOR_TYPE, getModelName, sceneImgDir } from '../../config/index'
-import { toPropertyDesc, storage, getCurrentPageParams, strUtil, checkInputNameIllegal } from '../../utils/index'
+import {
+  toPropertyDesc,
+  storage,
+  getCurrentPageParams,
+  strUtil,
+  checkInputNameIllegal,
+  emitter,
+} from '../../utils/index'
 import { adviceSceneNameList } from '../../config/scene'
 
 ComponentWithComputed({
@@ -28,7 +35,10 @@ ComponentWithComputed({
       (storage.get<number>('statusBarHeight') as number) +
       (storage.get<number>('navigationBarHeight') as number) +
       'px',
+    _sceneInfo: {} as Scene.SceneItem, //待编辑的一键场景
     _autosceneInfo: {} as AutoScene.AutoSceneItem, //待编辑的自动化
+    yijianSceneId: '',
+    isDefaultYijianScene: false, //是否是一键场景
     autoSceneId: '',
     sceneIcon: 'icon-1',
     sceneName: '',
@@ -180,11 +190,10 @@ ComponentWithComputed({
    */
   methods: {
     async onLoad() {
-      const { autosceneid, roomid } = getCurrentPageParams()
+      const { autosceneid, roomid, yijianSceneId } = getCurrentPageParams()
       console.log('接收到roomid', roomid)
-      this.setData({ autoSceneId: autosceneid, opearationType: 'auto' })
 
-      //处理三个传感器、场景和设备列表
+      // #region处理三个传感器、场景和设备列表
       await Promise.all([
         sceneStore.updateAllRoomSceneList(),
         deviceStore.updateAllRoomDeviceList(), //deviceStore.updateSubDeviceList(), //
@@ -206,8 +215,9 @@ ComponentWithComputed({
         ),
         sensorList,
       })
-
+      // #endregion 处理三个传感器、场景和设备列表
       if (autosceneid) {
+        this.setData({ autoSceneId: autosceneid, opearationType: 'auto' })
         const sensorlinkSelectList = [] as string[]
 
         const autoSceneInfo = autosceneStore.allRoomAutoSceneList.find(
@@ -353,8 +363,111 @@ ComponentWithComputed({
             this.updateSceneDeviceConditionsFlatten()
           },
         )
+        return
       }
+      if (yijianSceneId && roomid) {
+        this.setData({ yijianSceneId: yijianSceneId, opearationType: 'yijian' })
+        const sensorlinkSelectList = [] as string[]
 
+        // 获取当前scene
+        const sceneInfo = sceneStore.allRoomSceneList.find((item) => item.sceneId === yijianSceneId) as Scene.SceneItem
+        console.log('当前场景', sceneInfo)
+        this.data._sceneInfo = sceneInfo
+        this.setData({
+          sceneIcon: sceneInfo.sceneIcon,
+          sceneName: sceneInfo.sceneName,
+          roomId: sceneInfo.roomId,
+          isDefaultYijianScene: sceneInfo.isDefault === '1',
+        })
+
+        this.updateSceneDeviceConditionsFlatten()
+
+        //处理执行结果
+        const tempSceneDeviceActionsFlatten = [] as AutoScene.AutoSceneFlattenAction[]
+        const tempSceneDevicelinkSelectList: string[] = []
+
+        sceneInfo.deviceActions.forEach((action) => {
+          //设备
+          let deviceUniId = action.deviceId
+          if (action.proType === PRO_TYPE.switch) {
+            deviceUniId = `${action.deviceId}:${action.controlAction[0].modelName}`
+          }
+          let device = this.data.deviceList.find((item) => item.uniId === deviceUniId)
+          console.log('找到选项', device)
+          if (device) {
+            //是设备
+            if (device.proType === PRO_TYPE.switch) {
+              //是开关面板
+              action.controlAction.forEach((switchInPanel, switchIndex) => {
+                deviceUniId = `${action.deviceId}:${action.controlAction[switchIndex].modelName}`
+                device = this.data.deviceList.find((item) => item.uniId === deviceUniId)
+                if (device) {
+                  console.log('找到选项', device)
+                  const power = action.controlAction[switchIndex].power
+                  const desc = toPropertyDesc(device.proType, action.controlAction[switchIndex])
+                  tempSceneDevicelinkSelectList.push(device.uniId)
+                  tempSceneDeviceActionsFlatten.push({
+                    uniId: device.uniId,
+                    name: `${device.switchInfoDTOList[0].switchName} | ${device.deviceName}`,
+                    type: device.deviceType as 1 | 2 | 3 | 4 | 5 | 6,
+                    desc,
+                    pic: device.switchInfoDTOList[0].pic,
+                    proType: PRO_TYPE.switch,
+                    value: {
+                      modelName: action.controlAction[switchIndex].modelName,
+                      power,
+                    },
+                    orderNum: 0,
+                    dragId: device.uniId + Math.floor(Math.random() * 1001),
+                  })
+                  console.log(
+                    '初始化',
+                    switchInPanel,
+                    `${device.switchInfoDTOList[0].switchName} | ${device.deviceName}`,
+                  )
+                }
+              })
+            } else {
+              // const modelName = getModelName(device.proType, device.productId)
+              const property = {
+                // ...device.mzgdPropertyDTOList[modelName],
+                ...action.controlAction[0],
+              }
+              const desc = toPropertyDesc(device.proType, property)
+              tempSceneDevicelinkSelectList.push(device.uniId)
+              tempSceneDeviceActionsFlatten.push({
+                uniId: device.uniId,
+                name: device.deviceName,
+                type: device.deviceType as 1 | 2 | 3 | 4 | 5 | 6,
+                desc,
+                pic: device.pic as string,
+                proType: device.proType,
+                value: {
+                  ...property,
+                  modelName: getModelName(device.proType, device.productId),
+                },
+                orderNum: 0,
+                dragId: device.uniId + Math.floor(Math.random() * 1001),
+              })
+            }
+          } else {
+            console.log('设备不存在', action)
+          }
+        })
+        this.setData(
+          {
+            sensorlinkSelectList,
+            sceneDevicelinkSelectList: tempSceneDevicelinkSelectList,
+            sceneDeviceActionsFlatten: tempSceneDeviceActionsFlatten,
+          },
+          () => {
+            console.log('初始化已选择列表', this.data.sceneDevicelinkSelectList, this.data.sceneDeviceActionsFlatten)
+            this.updateSceneDeviceActionsFlatten()
+            this.updateSceneDeviceConditionsFlatten()
+          },
+        )
+        return
+      }
       if (roomid) {
         const deviceListInRoom: Device.DeviceItem[] = deviceStore.allRoomDeviceFlattenList.filter(
           (item) => item.roomId === roomid,
@@ -372,6 +485,7 @@ ComponentWithComputed({
             this.updateSceneDeviceConditionsFlatten()
           },
         )
+        return
       }
     },
 
@@ -647,6 +761,10 @@ ComponentWithComputed({
      */
     handleSelectCardShow() {
       // const switchUniId = this.data.checkedList[0]
+      // 默认场景不可编辑场景设备数据
+      if (this.data.isDefaultYijianScene) {
+        return
+      }
       if (
         this.data.opearationType === 'yijian' &&
         this.data.deviceList.filter((item) => !this.data.sceneDevicelinkSelectList.includes(item.uniId)).length === 0
@@ -970,6 +1088,10 @@ ComponentWithComputed({
           showTimeConditionPopup: true,
         })
       } else if (action.productId === 'touch') {
+        if (this.data.isDefaultYijianScene) {
+          // 默认情景不能换房间
+          return
+        }
         this.setData({
           showEditRoomPopup: true,
         })
@@ -990,6 +1112,10 @@ ComponentWithComputed({
      * @returns
      */
     handleAutoSceneActionEdit(e: { detail: number }) {
+      // 默认场景不可编辑场景设备数据
+      if (this.data.isDefaultYijianScene) {
+        return
+      }
       const index = e.detail
       const action = this.data.sceneDeviceActionsFlatten[index]
       console.log('handleAutoSceneActionEdit', action)
@@ -1095,6 +1221,10 @@ ComponentWithComputed({
     },
 
     async handleSave() {
+      if (this.data.opearationType === 'yijian' && this.data.yijianSceneId) {
+        this.updateYijianScene()
+        return
+      }
       if (this.data.opearationType === 'yijian') {
         this.go2dispatch()
         return
@@ -1370,6 +1500,68 @@ ComponentWithComputed({
         })
       }
     },
+    async updateYijianScene() {
+      if (this.data.sceneDeviceActionsFlatten.length === 0) {
+        // 删完actions按照删除场景处理
+        Dialog.confirm({
+          title: '清空操作将会删除场景，确定删除该场景？',
+        }).then(async () => {
+          const res = await deleteScene(this.data._sceneInfo.sceneId)
+          if (res.success) {
+            emitter.emit('sceneEdit')
+            homeStore.updateRoomCardList()
+            wx.navigateBack()
+          } else {
+            Toast({ message: '删除失败', zIndex: 9999 })
+          }
+        })
+        return
+      }
+
+      // 准备好数据内存
+      const data = {
+        sceneId: this.data._sceneInfo.sceneId,
+        updateType: '0',
+        conditionType: '0',
+        roomId: this.data.roomId,
+      } as Scene.UpdateSceneDto
+      // 检查场景名字是否变更
+      if (this.data.sceneName !== this.data._sceneInfo.sceneName) {
+        data.sceneName = this.data.sceneName
+      }
+      // 检查场景icon是否变更
+      if (this.data.sceneIcon !== this.data._sceneInfo.sceneIcon) {
+        data.sceneIcon = this.data.sceneIcon
+      }
+
+      if (this.data._isEditAction) {
+        data.deviceActions = []
+        data.updateType = data.updateType === '0' ? '1' : data.updateType === '2' ? '4' : '5'
+
+        // 场景动作数据统一在scene-request-list页面处理
+        storage.set('scene_data', data)
+        storage.set('sceneDeviceActionsFlatten', this.data.sceneDeviceActionsFlatten)
+
+        // 需要更新结果的情况，需要跳转页面等待上报结果
+        wx.redirectTo({
+          url: strUtil.getUrlWithParams('/package-automation/scene-request-list-yijian/index', {
+            sceneId: data.sceneId,
+          }),
+        })
+
+        return
+      }
+
+      const res = await updateScene(data)
+      if (res.success) {
+        emitter.emit('sceneEdit')
+        homeStore.updateRoomCardList()
+        Toast({ message: '修改成功', zIndex: 9999 })
+        wx.navigateBack()
+      } else {
+        Toast({ message: '修改失败', zIndex: 9999 })
+      }
+    },
     async go2dispatch() {
       // 检查场景名是否合法
       if (!this.data.sceneName) {
@@ -1422,21 +1614,44 @@ ComponentWithComputed({
       })
     },
     async handleAutoSceneDelete() {
-      const res = await Dialog.confirm({
-        title: '确定删除该自动化？',
-        zIndex: 9999,
-      }).catch(() => 'cancel')
+      if (this.data.autoSceneId) {
+        const res = await Dialog.confirm({
+          title: '确定删除该自动化？',
+          zIndex: 9999,
+        }).catch(() => 'cancel')
 
-      console.log('delAutoScene', res)
+        console.log('delAutoScene', res)
 
-      if (res === 'cancel') return
+        if (res === 'cancel') return
 
-      const delRes = await deleteScene(this.data.autoSceneId)
-      if (delRes.success) {
-        await autosceneStore.updateAllRoomAutoSceneList()
-        wx.navigateBack()
-      } else {
-        Toast({ message: '删除失败', zIndex: 9999 })
+        const delRes = await deleteScene(this.data.autoSceneId)
+        if (delRes.success) {
+          await autosceneStore.updateAllRoomAutoSceneList()
+          wx.navigateBack()
+        } else {
+          Toast({ message: '删除失败', zIndex: 9999 })
+        }
+        return
+      }
+      if (this.data.yijianSceneId) {
+        const res = await Dialog.confirm({
+          title: this.data.isDefaultYijianScene ? '默认场景删除后不可恢复，确定删除该默认场景？' : '确定删除该场景？',
+          zIndex: 9999,
+        }).catch(() => 'cancel')
+
+        console.log('delAutoScene', res)
+
+        if (res === 'cancel') return
+
+        const delRes = await deleteScene(this.data._sceneInfo.sceneId)
+        if (delRes.success) {
+          emitter.emit('sceneEdit')
+          homeStore.updateRoomCardList()
+          wx.navigateBack()
+        } else {
+          Toast({ message: '删除失败', zIndex: 9999 })
+        }
+        return
       }
     },
     /* 执行结果拖拽相关方法 start */
